rerouting nav component for user friendly
