+++ Client/src/main.tsx
@@ -2,16 +2,18 @@ import ReactDOM from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import './utils/index.css';
 
+import Login from './pages/Home.tsx'
 import App from './App.tsx';
 import ErrorPage from './pages/ErrorPage.tsx';
-import Home from './pages/Home.tsx';
+import Shop from './pages/Shop.tsx';
 import Register from './pages/Register.tsx'
-import Login from './pages/Login.tsx'
 import Details from './pages/Details.tsx'
 import Checkout from './pages/Checkout.tsx'
 import Cart from './pages/Cart.tsx'
+
 import { CartProvider } from './components/CartState.tsx';
 import { WishlistProvider } from './components/wishlistSate.tsx';
+
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'bootstrap/dist/js/bootstrap.bundle.min.js';
 
@@ -30,15 +32,15 @@ const router = createBrowserRouter([
     children: [
       {
         index: true,
-        element: <Home />
+        element: <Login />
       }, 
       {
         path: '/Register',
         element: <Register />
       }, 
       {
-        path: '/Login',
-        element: <Login />
+        path: '/Shop',
+        element: <Shop />
       }, 
       {
         path: '/Details',
